+++ src/doom/statdump.cpp
@@ -286,8 +286,8 @@ static void PrintStats(FILE *stream, const wbstartstruct_t *stats)
     PrintLevelName(stream, stats->epsd, stats->last);
     fprintf(stream, "\n");
 
-    leveltime = stats->plyr[0].stime / TICRATE;
-    partime   = stats->partime / TICRATE;
+    leveltime = static_cast<short>(stats->plyr[0].stime / TICRATE);
+    partime   = static_cast<short>(stats->partime / TICRATE);
     fprintf(stream, "Time: %i:%02i", leveltime / 60, leveltime % 60);
     fprintf(stream, " (par: %i:%02i)\n", partime / 60, partime % 60);
     fprintf(stream, "\n");
